+++ src/layouts/SidebarLayout/Sidebar/index.js
@@ -65,7 +65,7 @@ function Sidebar() {
           </Box>
           <Divider
             sx={{
-              mt: theme.spacing(3),
+              mt: theme.spacing(1),
               mx: theme.spacing(2),
               background: theme.colors.alpha.trueWhite[10]
             }}